requirements.txt
spacy
pytest
pydantic
google-cloud-firestore
fastapi
uvicorn